- 21-02-2005 fixed: iTunes sharing caused crash if no connection (or iTunes sharing) was available.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@2889 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/FileSystem/DAAPDirectory.cpp
@@ -116,7 +116,7 @@ bool  CDAAPDirectory::GetDirectory(const CStdString& strPath, VECFILEITEMS &item
 			return false;
 		}
 
-		if (DAAP_ClientHost_Connect(m_thisHost) < 0)
+		if ((int)DAAP_ClientHost_Connect(m_thisHost) < 0)
 		{
 			CloseDAAP();
 			return false;
